#200 - 金額の単位を設定可能にする
- 幾つかの金額に、デフォルトに設定したアセットの単位(前置/後置)を表示するように変更
- グラフの縦軸の表記にデフォルトに設定したアセットの名称を表示するように変更

+++ HouseholdAccountBook/Infrastructure/CSV/CSVFileDao.cs
@@ -68,7 +68,7 @@ namespace HouseholdAccountBook.Infrastructure.CSV
                                 name = null;
                             }
                             if (!reader.TryGetField(expensesIndex - 1, out string valueStr) ||
-                                !int.TryParse(valueStr, NumberStyles.Any, NumberFormatInfo.CurrentInfo, out int value)) {
+                                !decimal.TryParse(valueStr, NumberStyles.Any, NumberFormatInfo.CurrentInfo, out decimal value)) {
                                 continue;
                             }
